Move test_reduction to directory
